HomeWork

+++ src/test/java/CalculatorTest.java
@@ -7,15 +7,15 @@ public class CalculatorTest {
 
     @Test
     void averageValueArrayTest(){
-        float[] array = new float[]{1, 2, 3};
+        int[] array = new int[]{1, 2, 3};
 
         assertEquals(Calculator.averageValueArray(array), 2);
     }
 
     @Test
     void comparisonAverageValueTestLeftMoreRight(){
-        float[] arrrayLeft = new float[]{ 10, 30, 40, 50};
-        float[] arrrayRight = new float[]{ 1, 3, 4, 5};
+        int[] arrrayLeft = new int[]{ 10, 30, 40, 50};
+        int[] arrrayRight = new int[]{ 1, 3, 4, 5};
 
         assertEquals(Calculator.comparisonAverageValueArray(arrrayLeft, arrrayRight), "Первый список имеет большее среднее значение");
 
@@ -24,8 +24,8 @@ public class CalculatorTest {
     @Test
     void comparisonAverageValueTestLeftLessRight(){
 
-        float[] arrrayLeft = new float[]{ 1, 3, 4, 5};
-        float[] arrrayRight = new float[]{ 10, 30, 40, 50};
+        int[] arrrayLeft = new int[]{ 1, 3, 4, 5};
+        int[] arrrayRight = new int[]{ 10, 30, 40, 50};
 
         assertEquals(Calculator.comparisonAverageValueArray(arrrayLeft, arrrayRight), "Второй список имеет большее среднее значение");
     }
@@ -33,8 +33,8 @@ public class CalculatorTest {
     @Test
     void comparisonAverageValueTestLeftEqualRight(){
 
-        float[] arrrayLeft = new float[]{ 10, 30, 40, 50};
-        float[] arrrayRight = new float[]{ 10, 30, 40, 50};
+        int[] arrrayLeft = new int[]{ 10, 30, 40, 50};
+        int[] arrrayRight = new int[]{ 10, 30, 40, 50};
 
         assertEquals(Calculator.comparisonAverageValueArray(arrrayLeft, arrrayRight), "Средние значения равны");
     }
